[C++][Graph] LC 684. Redundant Connection
This might not be solved using simple DFS for cycle find since there can
be case where lets say we start on root 'N' and one of its child makes a cycle
with its own nbr but root N is not part of it but it would be present in DFS visited
stack and cause wrong answer. Hence we probably need to start DFS from each node
(not just any e.g 1) and see which smallest visited stack is containing cycle.
Testcases to watch out for:
TC 1: [[1,3],[3,4],[1,5],[3,5],[2,3]]
TC 2: [[2,7],[7,8],[3,6],[2,5],[6,8],[4,8],[2,8],[1,8],[7,10],[3,9]]